+++ spec/fixtures/non-zero.js
@@ -1,0 +1,3 @@
+'use strict'
+
+process.exitCode = 2